refine waste reminder module split
